More bugfixes.
* examples/mblconv.c (main): Return 0.
* examples/murun.c (main): Fix iteration body.
* imap4d/append.c: Do not pass constructed string as format
argument to io_completion_response, use %s instead.
* imap4d/close.c: Likewise.
* imap4d/create.c: Likewise.
* imap4d/delete.c: Likewise.
* imap4d/rename.c: Likewise.
* imap4d/status.c: Likewise.
* libmailutils/base/mutil.c (mutil_parse_field_map): Fix return
code in case of wordsplit failure.
* libmailutils/cfg/parser.y (GETSNUM): Initialize sign.
* libmailutils/stream/streamref.c (_streamref_ctl)
<MU_IOCTL_SEEK_LIMITS>: Initialize lim.
* libmailutils/string/wordsplit.c: Typecast format width arguments
to int.
* libmu_sieve/extensions/vacation.c (check_db): Initialize days.
Pass correct time value to test_and_update_prop.
* libproto/pop/mbox.c (pop_create_pop3_message): Fix return code.
* mail/mail.c (do_and_quit): Pass the command to util_do_command
via the %s format.
* mail/unset.c (mail_unset): Likewise.
* mh/mh_fmtgram.y (function production): LIkewise for mu_error.
* mu/logger.c (mutool_logger): Fix return code in case
of mu_stdstream_strerr_create failure.

+++ libmailutils/string/wordsplit.c
@@ -412,7 +412,7 @@ mu_wordsplit_dump_nodes (struct mu_wordsplit *wsp)
       else
 	wsp->ws_debug ("%4d: %p: %#04x (%s):%.*s;",
 		       n, p, p->flags, wsnode_flagstr (p->flags),
-		       p->v.segm.end - p->v.segm.beg,
+		       (int)(p->v.segm.end - p->v.segm.beg),
 		       wsp->ws_input + p->v.segm.beg);
     }
 }
@@ -773,7 +773,7 @@ expvar (struct mu_wordsplit *wsp, const char *str, size_t len,
   else
     {
       if (wsp->ws_flags & MU_WRDSF_WARNUNDEF)
-	wsp->ws_error (_("warning: undefined variable `%.*s'"), i, str);
+	wsp->ws_error (_("warning: undefined variable `%.*s'"), (int) i, str);
       if (wsp->ws_flags & MU_WRDSF_KEEPUNDEF)
 	value = NULL;
       else
@@ -1346,7 +1346,7 @@ mu_wordsplit_len (const char *command, size_t len, struct mu_wordsplit *wsp,
     return rc;
 
   if (wsp->ws_flags & MU_WRDSF_SHOWDBG)
-    wsp->ws_debug ("Input:%.*s;", len, command);
+    wsp->ws_debug ("Input:%.*s;", (int)len, command);
 
   if (wsp->ws_flags & MU_WRDSF_NOSPLIT)
     {
